feat: Phase 2 audio + val leakage fix + progressive plot + UX polish
This is a coherent end-to-end batch wiring audio training, fixing two
serious correctness bugs caught while testing on a real Kaggle dataset,
and polishing several UX rough edges.

Phase 2 Audio foundations:
- AudioDatasetEntry + registry plumbing mirrors ImageDatasetEntry
  (folder_path, feature_type, SR/n_fft/hop/n_mels/n_mfcc/max_duration,
  probed feature_rows/feature_cols, class_names). ClearAllTabularDatasets
  now wipes audio entries too; UnregisterNodeDatasetIfOwned covers audio.
- AudioDatasetBatcher is a new IBatcher with shuffled split, per-sample
  z-score (mel dB range would saturate Xavier-init Dense layers without
  it), and probed feature shape read from the first valid sample.
- TrainingManager::StartTrainingAudio mirrors StartTrainingImage.
- StartTrainingFromGraph adds the IsAudioDataset dispatch branch.
- DataInputDialog audio category is no longer a "not yet supported"
  gate — has a folder picker, ClassSubdirs / FlatWithCSV layout combo,
  optional labels CSV with auto-detected delimiter + filename/label
  columns, sample rate + max duration inputs. Apply path builds the
  registry entry and runs a probe load to validate folder structure.
- graph_compiler audio domain check: file_category="audio" sets
  preprocessing_domain=Audio and skips the misleading "no label column"
  warning (labels come from folder structure or CSV mapping).
- AudioDataset supports a new FlatWithCSV mode via LoadFromCSV: auto
  delimiter detection (comma/semicolon/tab), header auto-detection,
  filename/label column auto-detection with priority lists, stem-based
  filename resolution so "call_id" without .flac matches call_id.flac.
- Three cyxgraph examples: drone_sound_detector, call_center_sentiment,
  speech_command_classifier. Drone defaults to 1s max_duration to match
  actual clip length (was 2s, 5x faster training).

Val pipeline leakage fix (was giving fake 100% metrics):
- Before: image/audio batchers used sequential "first 80%" split and
  active_train_ibatcher = active_val_ibatcher = same pointer. Val pass
  iterated training data with dropout off — not real validation.
  Paired with alphabetical class ordering in ScanDirectory, val ended
  up 100% one class, and a "always predict class N" model scored 100%.
- Now: both batchers shuffle full index list then split into
  train_indices_ / val_indices_. IBatcher adds a SetPhase virtual;
  image and audio override to swap which index set Reset() uses.
  training_executor calls SetPhase(Val) before RunValidationArrow and
  SetPhase(Train) after.

Silent-audio detection (robustness fix):
- AudioDataset::ExtractFeatures rejects files with all-zero samples
  (returns invalid with "audio file contains only silence"). Caught
  1089 / 11704 silent files in the Kaggle Binary Drone Audio dataset
  that would otherwise train as garbage without any warning.
- GetItem rate-limits the warning spam: first 10 full lines, then
  every 100th, via std::atomic counter.
- The probe loop in AudioDataset constructor now tries up to 32
  samples before giving up, so datasets with silent files near the
  top of the alphabetical scan still probe the feature shape
  correctly instead of falling back to the wrong closed-form formula.

Memory tab restore fix:
- AudioDatasetEntry persists probed feature_rows/feature_cols so
  dialog reopen computes the "if fully cached" size estimate without
  re-probing. Image restore branch now also computes the estimate
  (224x224x3 default) instead of showing 0 B. New loaded_memory_is_estimate_
  flag so the Memory tab labels lazy-loaded datasets as
  "Lazy-loaded - ~X if fully cached (estimated)" instead of the
  misleading "In Memory - 0 B". file_category, audio_layout,
  audio_labels_csv, audio_filename_col, audio_label_col are now
  persisted in node params so dialog reopen restores them.

Progressive plot fix (training loss draws during long epochs):
- AddLossPoint / AddAccuracyPoint signature int -> double so batch
  callbacks can push fractional-epoch x-axis points.
- training_manager batch_callback adds a loss point per batch with
  x = (epoch - 1) + batch / total_batches. Audio epochs at 8 min/epoch
  no longer show an empty plot while waiting for the first epoch to
  finish — train curve fills in smoothly.

tofix.md updates: added the big "TrainingExecutor walks pin connections"
deferred item (the executor still reads from registry by name, not
from the graph), the sub-epoch validation enhancement, and the new
cross-domain "Dataset sanity audit at Apply time" section explaining
why every domain needs quality checks (integration tests on real
datasets catch what unit tests can't — the drone silent files being
Exhibit A). Marked resolved items.

Also: ORGANIZATION_PROFILE.md website link updated to
https://web-psi-drab.vercel.app/

+++ cyxwiz-engine/src/gui/panels/training_plot_panel.cpp
@@ -123,7 +123,7 @@ void TrainingPlotPanel::Render() {
     ImGui::End();
 }
 
-void TrainingPlotPanel::AddLossPoint(int epoch, double train_loss, double val_loss) {
+void TrainingPlotPanel::AddLossPoint(double epoch, double train_loss, double val_loss) {
     std::lock_guard<std::mutex> lock(data_mutex_);
 
     train_loss_.epochs.push_back(epoch);
@@ -137,7 +137,7 @@ void TrainingPlotPanel::AddLossPoint(int epoch, double train_loss, double val_lo
     }
 }
 
-void TrainingPlotPanel::AddAccuracyPoint(int epoch, double train_acc, double val_acc) {
+void TrainingPlotPanel::AddAccuracyPoint(double epoch, double train_acc, double val_acc) {
     std::lock_guard<std::mutex> lock(data_mutex_);
 
     train_accuracy_.epochs.push_back(epoch);
